requirements.txt
streamlit
pytesseract
Pillow
openai
langchain
python-dotenv
